Исправлены ошибки при добавлении и удалении контрактов

+++ Scanning/DeleteContractCommand.cs
@@ -6,28 +6,31 @@ using WpfApp.Service;
 
 namespace WpfApp.Scanning
 {
+	/// <summary>
+	/// Команда удаления документа при сканировании
+	/// </summary>
 	public class DeleteContractCommand : IScanningCommand
 	{
 		public bool IsWorking { get; private set; }
 
-		private string _boxId;
-
-		public DeleteContractCommand(string boxId)
+		public DeleteContractCommand()
 		{
-			_boxId = boxId;
 			IsWorking = false;
 		}
 
-		public async void DoWork()
+		public void DoWork(string barCode)
 		{
 			IsWorking = true;
 			var contracts = Contract.Repository.GetAll().ToList();
 
-			var random = new Random();
-			var randomContract = contracts[random.Next(0, contracts.Count - 1)];
-			ConsoleWriter.Write($"Договор с номером {randomContract.Number} удален");
-			Contract.Repository.DeleteById(randomContract.Id);
-			Thread.Sleep(2000);
+			if (contracts.Any())
+			{
+				var random = new Random();
+				var randomContract = contracts[random.Next(0, contracts.Count - 1)];
+				ConsoleWriter.Write($"Договор с номером {randomContract.Number} удален");
+				Contract.Repository.DeleteById(randomContract.Id);
+			}
+
 			IsWorking = false;
 		}
 	}
